requirements.txt
fastapi
uvicorn
sqlalchemy
aiosqlite
Jinja2
python-multipart
pydantic_settings
psycopg2
jose
passlib
fastapi-users
asyncpg
fastapi-users[sqlalchemy]
python-jose[cryptography]
passlib[bcrypt]
itsdangerous
pandas
openpyxl
xlrd==1.2.0
ujson
aiofiles